+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+* Dedent current line on `else`, `elif`, `except`, and `finally` statements
 * Don't dedent on keywords if they appear in triple quoted string.
 
 ### 0.8.1